Bug 960653 - part 2 - change JSJitInfo.{protoID,depth} to 16-bit fields; r=bz,efaust
UltraBlame original commit: 4b70796946456df101de34d7b4ef84619fe58f08

+++ js/src/jsfriendapi.h
@@ -1517,8 +1517,8 @@ struct JSJitInfo {
         JSParallelNative parallelNative;
     };
 
-    uint32_t protoID;
-    uint32_t depth;
+    uint16_t protoID;
+    uint16_t depth;
